Updated JFrame. Still needs to implement a TrackMate macro
